refactor: 주문 테이블이 테이블 그룹을 간접 참조하도록 수정
