--- src/main/java/kitchenpos/application/TableGroupService.java
@@ -1,59 +1,0 @@
-package kitchenpos.application;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-import kitchenpos.domain.order.OrderRepository;
-import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.ordertable.OrderTable;
-import kitchenpos.domain.ordertable.OrderTableRepository;
-import kitchenpos.domain.ordertable.TableGroup;
-import kitchenpos.domain.ordertable.TableGroupRepository;
-import kitchenpos.dto.request.OrderTableIdRequest;
-import kitchenpos.dto.request.TableGroupRequest;
-import kitchenpos.dto.response.TableGroupResponse;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional(readOnly = true)
-public class TableGroupService {
-    private final OrderRepository orderRepository;
-    private final OrderTableRepository orderTableRepository;
-    private final TableGroupRepository tableGroupRepository;
-
-    public TableGroupService(final OrderRepository orderRepository, final OrderTableRepository orderTableRepository,
-                             final TableGroupRepository tableGroupRepository) {
-        this.orderRepository = orderRepository;
-        this.orderTableRepository = orderTableRepository;
-        this.tableGroupRepository = tableGroupRepository;
-    }
-
-    @Transactional
-    public TableGroupResponse create(final TableGroupRequest request) {
-        List<Long> orderTableIds = request.getOrderTables().stream()
-                .map(OrderTableIdRequest::getId)
-                .collect(Collectors.toList());
-        List<OrderTable> orderTables = orderTableRepository.findAllByIdIn(orderTableIds);
-
-        TableGroup tableGroup = new TableGroup(orderTables);
-        return new TableGroupResponse(tableGroupRepository.save(tableGroup));
-    }
-
-    @Transactional
-    public void ungroup(final Long tableGroupId) {
-        TableGroup tableGroup = tableGroupRepository.findById(tableGroupId)
-                .orElseThrow(IllegalArgumentException::new);
-
-        List<Long> orderTableIds = tableGroup.getOrderTables().stream()
-                .map(OrderTable::getId)
-                .collect(Collectors.toList());
-
-        if (orderRepository.existsByOrderTableIdInAndOrderStatusIn(
-                orderTableIds, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
-            throw new IllegalArgumentException("테이블의 주문이 완료되지 않아 단체 지정을 취소할 수 없습니다.");
-        }
-
-        tableGroup.ungroup();
-    }
-}
