+++ src/main/java/kitchenpos/application/TableService.java
@@ -7,20 +7,29 @@ import kitchenpos.domain.order.OrderRepository;
 import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.ordertable.OrderTable;
 import kitchenpos.domain.ordertable.OrderTableRepository;
+import kitchenpos.domain.tablegroup.TableGroup;
+import kitchenpos.domain.tablegroup.TableGroupRepository;
+import kitchenpos.dto.request.OrderTableIdRequest;
 import kitchenpos.dto.request.OrderTableRequest;
+import kitchenpos.dto.request.TableGroupRequest;
 import kitchenpos.dto.response.OrderTableResponse;
+import kitchenpos.dto.response.TableGroupResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @Transactional(readOnly = true)
 public class TableService {
-    private final OrderRepository orderRepository;
     private final OrderTableRepository orderTableRepository;
+    private final TableGroupRepository tableGroupRepository;
+    private final OrderRepository orderRepository;
 
-    public TableService(final OrderRepository orderRepository, final OrderTableRepository orderTableRepository) {
-        this.orderRepository = orderRepository;
+    public TableService(final OrderTableRepository orderTableRepository,
+                        final TableGroupRepository tableGroupRepository,
+                        final OrderRepository orderRepository) {
         this.orderTableRepository = orderTableRepository;
+        this.tableGroupRepository = tableGroupRepository;
+        this.orderRepository = orderRepository;
     }
 
     @Transactional
@@ -56,5 +65,33 @@ public class TableService {
 
         savedOrderTable.changeNumberOfGuests(request.getNumberOfGuests());
         return new OrderTableResponse(savedOrderTable);
+    }
+
+    @Transactional
+    public TableGroupResponse group(final TableGroupRequest request) {
+        List<Long> orderTableIds = request.getOrderTables().stream()
+                .map(OrderTableIdRequest::getId)
+                .collect(Collectors.toList());
+        List<OrderTable> orderTables = orderTableRepository.findAllByIdIn(orderTableIds);
+
+        TableGroup tableGroup = new TableGroup(orderTables);
+        return new TableGroupResponse(tableGroupRepository.save(tableGroup));
+    }
+
+    @Transactional
+    public void ungroup(final Long tableGroupId) {
+        TableGroup tableGroup = tableGroupRepository.findById(tableGroupId)
+                .orElseThrow(IllegalArgumentException::new);
+
+        List<Long> orderTableIds = tableGroup.getOrderTables().stream()
+                .map(OrderTable::getId)
+                .collect(Collectors.toList());
+
+        if (orderRepository.existsByOrderTableIdInAndOrderStatusIn(
+                orderTableIds, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
+            throw new IllegalArgumentException("테이블의 주문이 완료되지 않아 단체 지정을 취소할 수 없습니다.");
+        }
+
+        tableGroup.ungroup();
     }
 }
