+++ src/main/java/kitchenpos/domain/tablegroup/TableGroup.java
@@ -1,8 +1,7 @@
-package kitchenpos.domain.ordertable;
+package kitchenpos.domain.tablegroup;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Objects;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -11,7 +10,9 @@ import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
 import javax.persistence.OneToMany;
+import kitchenpos.domain.ordertable.OrderTable;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
@@ -27,7 +28,8 @@ public class TableGroup {
     @Column(name = "createdDate")
     private LocalDateTime createdDate;
 
-    @OneToMany(cascade = CascadeType.PERSIST, fetch = FetchType.LAZY, mappedBy = "tableGroup")
+    @OneToMany(cascade = CascadeType.PERSIST, fetch = FetchType.LAZY)
+    @JoinColumn(name = "table_group_id")
     private List<OrderTable> orderTables;
 
     protected TableGroup() {
@@ -37,25 +39,13 @@ public class TableGroup {
         validateOrderTables(orderTables);
         this.orderTables = orderTables;
         for (OrderTable orderTable : orderTables) {
-            orderTable.changeTableGroup(this);
+            orderTable.group(id);
         }
     }
 
     private void validateOrderTables(final List<OrderTable> orderTables) {
         if (orderTables.size() < 2) {
             throw new IllegalArgumentException("그룹으로 지정할 테이블은 2개 이상이어야 합니다.");
-        }
-        for (OrderTable orderTable : orderTables) {
-            validateOrderTable(orderTable);
-        }
-    }
-
-    private void validateOrderTable(final OrderTable orderTable) {
-        if (!orderTable.isEmpty()) {
-            throw new IllegalArgumentException("그룹으로 지정할 테이블이 비어있지 않습니다.");
-        }
-        if (Objects.nonNull(orderTable.getTableGroup())) {
-            throw new IllegalArgumentException("그룹으로 지정할 테이블이 이미 그룹으로 지정되어 있습니다.");
         }
     }
