+++ src/main/java/kitchenpos/dto/response/OrderTableResponse.java
@@ -1,8 +1,6 @@
 package kitchenpos.dto.response;
 
-import java.util.Objects;
 import kitchenpos.domain.ordertable.OrderTable;
-import kitchenpos.domain.ordertable.TableGroup;
 
 public class OrderTableResponse {
     private Long id;
@@ -14,15 +12,7 @@ public class OrderTableResponse {
     }
 
     public OrderTableResponse(final OrderTable orderTable) {
-        this(orderTable.getId(), getTableGroupId(orderTable), orderTable.getNumberOfGuests(), orderTable.isEmpty());
-    }
-
-    private static Long getTableGroupId(final OrderTable orderTable) {
-        TableGroup tableGroup = orderTable.getTableGroup();
-        if (Objects.isNull(tableGroup)) {
-            return null;
-        }
-        return tableGroup.getId();
+        this(orderTable.getId(), orderTable.getTableGroupId(), orderTable.getNumberOfGuests(), orderTable.isEmpty());
     }
 
     public OrderTableResponse(final Long id, final Long tableGroupId, final int numberOfGuests, final boolean empty) {
