--- src/main/java/kitchenpos/ui/TableGroupRestController.java
@@ -1,36 +1,0 @@
-package kitchenpos.ui;
-
-import java.net.URI;
-import kitchenpos.application.TableGroupService;
-import kitchenpos.dto.request.TableGroupRequest;
-import kitchenpos.dto.response.TableGroupResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequestMapping("/api/table-groups")
-@RestController
-public class TableGroupRestController {
-    private final TableGroupService tableGroupService;
-
-    public TableGroupRestController(final TableGroupService tableGroupService) {
-        this.tableGroupService = tableGroupService;
-    }
-
-    @PostMapping
-    public ResponseEntity<TableGroupResponse> create(@RequestBody final TableGroupRequest request) {
-        TableGroupResponse response = tableGroupService.create(request);
-        URI uri = URI.create("/api/table-groups/" + response.getId());
-        return ResponseEntity.created(uri).body(response);
-    }
-
-    @DeleteMapping("/{tableGroupId}")
-    public ResponseEntity<Void> ungroup(@PathVariable final Long tableGroupId) {
-        tableGroupService.ungroup(tableGroupId);
-        return ResponseEntity.noContent().build();
-    }
-}
