+++ src/test/java/kitchenpos/domain/ordertable/OrderTableTest.java
@@ -13,7 +13,7 @@ class OrderTableTest {
 
     @Test
     void 그룹화된_테이블은_빈_상태로_변경할_수_없다() {
-        OrderTable orderTable = new OrderTable(new TableGroup(), 2, false);
+        OrderTable orderTable = new OrderTable(1L, 2, false);
 
         assertThatThrownBy(() -> orderTable.changeEmpty(true))
                 .isInstanceOf(IllegalArgumentException.class);
