+++ src/test/java/kitchenpos/domain/tablegroup/TableGroupTest.java
@@ -1,9 +1,10 @@
-package kitchenpos.domain.ordertable;
+package kitchenpos.domain.tablegroup;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import java.util.List;
+import kitchenpos.domain.ordertable.OrderTable;
 import org.junit.jupiter.api.Test;
 
 class TableGroupTest {
@@ -34,12 +35,8 @@ class TableGroupTest {
 
     @Test
     void 이미_그룹으로_지정된_테이블은_그룹으로_지정할_수_없다() {
-        OrderTable orderTable1 = new OrderTable(null, 2, true);
-        OrderTable orderTable2 = new OrderTable(null, 2, true);
-        TableGroup tableGroup = new TableGroup(List.of(orderTable1, orderTable2));
-
-        OrderTable alreadyGroupedOrderTable1 = new OrderTable(tableGroup, 2, true);
-        OrderTable alreadyGroupedOrderTable2 = new OrderTable(tableGroup, 2, true);
+        OrderTable alreadyGroupedOrderTable1 = new OrderTable(1L, 2, true);
+        OrderTable alreadyGroupedOrderTable2 = new OrderTable(1L, 2, true);
 
         assertThatThrownBy(() -> new TableGroup(List.of(alreadyGroupedOrderTable1, alreadyGroupedOrderTable2)))
                 .isInstanceOf(IllegalArgumentException.class);
